refactor: move and update variants to src/lib/variants.ts
Moves the style variants definitions to src/lib/variants.ts to follow standard library conventions. Updates imports across all components to use the aliased path @/lib/variants. Fixes the internal import in the variants file itself to use @/styles/design-tokens. Ensure TypeScript type check passes.
